Lab1, Planet.java are Done.
More info:-
I have justed forked the skeleton of cs61b for 2018 and moved the
files from the old repo. to this.
This repo. has lab1 files completed and Planet.java. Moreover,
The problems upto 'Drawing the Background' have passed the tests.

Note: This fork was made because some of the files (e.g images) did
not come to my local machine.